// apps/admin/src/api/generated/schemas/responseEnvelopeRefreshTokenResponse.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * x-ear API
 * Backend API for x-ear CRM
 * OpenAPI spec version: 1.0.0
 */
import type { ResponseEnvelopeRefreshTokenResponseData } from './responseEnvelopeRefreshTokenResponseData';
import type { ResponseEnvelopeRefreshTokenResponseMessage } from './responseEnvelopeRefreshTokenResponseMessage';
import type { ResponseEnvelopeRefreshTokenResponseError } from './responseEnvelopeRefreshTokenResponseError';
import type { ResponseEnvelopeRefreshTokenResponseMeta } from './responseEnvelopeRefreshTokenResponseMeta';
import type { ResponseEnvelopeRefreshTokenResponseRequestId } from './responseEnvelopeRefreshTokenResponseRequestId';

export interface ResponseEnvelopeRefreshTokenResponse {
  success?: boolean;
  data?: ResponseEnvelopeRefreshTokenResponseData;
  message?: ResponseEnvelopeRefreshTokenResponseMessage;
  error?: ResponseEnvelopeRefreshTokenResponseError;
  meta?: ResponseEnvelopeRefreshTokenResponseMeta;
  requestId?: ResponseEnvelopeRefreshTokenResponseRequestId;
  timestamp?: string;
}